modernize layout and move to go modules
