+++ Makefile
@@ -4,12 +4,10 @@ BINARY_SERVER   = homer-server
 DATE    ?= $(shell date +%FT%T%z)
 VERSION ?= $(shell git describe --tags --always --dirty --match=v* 2> /dev/null || \
 	    cat $(CURDIR)/.version 2> /dev/null || echo v0)
-GOPATH   = $(CURDIR)/.gopath~
+#GOPATH   = $(CURDIR)/.gopath~
 BIN      = $(GOPATH)/bin
 BASE     = $(GOPATH)/src/$(PACKAGE)
 PKGS     = $(or $(PKG),$(shell cd $(BASE) && env GOPATH=$(GOPATH) $(GO) list ./... | grep -v "^$(PACKAGE)/vendor/"))
-# ignore indefatigable (generated code) from linting
-LINT_PKGS = $(or $(PKG),$(shell cd $(BASE) && env GOPATH=$(GOPATH) $(GO) list ./... | grep -v "^$(PACKAGE)/vendor/"))
 TESTPKGS = $(shell env GOPATH=$(GOPATH) $(GO) list -f '{{ if or .TestGoFiles .XTestGoFiles }}{{ .ImportPath }}{{ end }}' $(PKGS))
 GO      = go
 GODOC   = godoc
@@ -21,18 +19,19 @@ M = $(shell printf "\033[34;1m✈\033[0m")
 
 .PHONY: all
 all: client server
+
 .PHONY: client
-client: fmt lint vendor | $(BASE) ; $(info $(M) building client executable…) @ ## Build program binary
+client: fmt | $(BASE) ; $(info $(M) building client executable…) @ ## Build program binary
 	$Q cd $(BASE) && $(GO) build \
 	  -tags release \
-	  -ldflags '-X $(PACKAGE)/version.Version=$(VERSION) -X $(PACKAGE)/version.BuildDate=$(DATE) -X $(PACKAGE)/version.Package=$(PACKAGE)' \
-	  -o bin/$(BINARY_CLIENT) client.go \
+	  -ldflags '-X $(PACKAGE)/internal/version.Version=$(VERSION) -X $(PACKAGE)/internal/version.BuildDate=$(DATE) -X $(PACKAGE)/internal/version.Package=$(PACKAGE)' \
+	  -o bin/$(BINARY_CLIENT) ./cmd/client \
 	  && echo "Built bin/$(BINARY_CLIENT): $(VERSION) $(DATE)"
-server: fmt lint vendor | $(BASE) ; $(info $(M) building server executable…) @ ## Build program binary
+server: fmt | $(BASE) ; $(info $(M) building server executable…) @ ## Build program binary
 	$Q cd $(BASE) && $(GO) build \
 	  -tags release \
-	  -ldflags '-X $(PACKAGE)/version.Version=$(VERSION) -X $(PACKAGE)/version.BuildDate=$(DATE) -X $(PACKAGE)/version.Package=$(PACKAGE)' \
-	  -o bin/$(BINARY_SERVER) server.go \
+	  -ldflags '-X $(PACKAGE)/internal/version.Version=$(VERSION) -X $(PACKAGE)/internal/version.BuildDate=$(DATE) -X $(PACKAGE)/internal/version.Package=$(PACKAGE)' \
+	  -o bin/$(BINARY_SERVER) ./cmd/server \
 	  && echo "Built bin/$(BINARY_SERVER): $(VERSION) $(DATE)"
 .PHONY: release
 release: export GOOS=linux
@@ -52,37 +51,6 @@ deploy-release: release
 	scp bin/$(BINARY)-$(VERSION) $(FSHOST):$(FSPATH)/
 	$Q echo Deployed release to https://$(FSHOST)/$(FSWWWPATH)/$(BINARY)-$(VERSION)
 
-$(BASE): ; $(info $(M) setting GOPATH…)
-	@mkdir -p $(dir $@)
-	@ln -sf $(CURDIR) $@
-
-# Tools
-
-GODEP = $(BIN)/dep
-$(BIN)/dep: | $(BASE) ; $(info $(M) building go dep…)
-	$(GO) get github.com/golang/dep/cmd/dep
-	echo $(GOPATH)
-
-GOLINT = $(BIN)/golint
-$(BIN)/golint: | $(BASE) ; $(info $(M) building golint…)
-	$Q $(GO) get github.com/golang/lint/golint
-
-GOCOVMERGE = $(BIN)/gocovmerge
-$(BIN)/gocovmerge: | $(BASE) ; $(info $(M) building gocovmerge…)
-	$Q $(GO) get github.com/wadey/gocovmerge
-
-GOCOV = $(BIN)/gocov
-$(BIN)/gocov: | $(BASE) ; $(info $(M) building gocov…)
-	$Q $(GO) get github.com/axw/gocov/...
-
-GOCOVXML = $(BIN)/gocov-xml
-$(BIN)/gocov-xml: | $(BASE) ; $(info $(M) building gocov-xml…)
-	$Q $(GO) get github.com/AlekSi/gocov-xml
-
-GO2XUNIT = $(BIN)/go2xunit
-$(BIN)/go2xunit: | $(BASE) ; $(info $(M) building go2xunit…)
-	$Q $(GO) get github.com/tebeka/go2xunit
-
 # Tests
 
 TEST_TARGETS := test-default test-bench test-short test-verbose test-race
@@ -93,10 +61,10 @@ test-verbose: ARGS=-v            ## Run tests in verbose mode with coverage repo
 test-race:    ARGS=-race         ## Run tests with race detector
 $(TEST_TARGETS): NAME=$(MAKECMDGOALS:test-%=%)
 $(TEST_TARGETS): test
-check test tests: fmt lint vendor | $(BASE) ; $(info $(M) running $(NAME:%=% )tests…) @ ## Run tests
+check test tests: fmt lint | $(BASE) ; $(info $(M) running $(NAME:%=% )tests…) @ ## Run tests
 	$Q cd $(BASE) && $(GO) test -timeout $(TIMEOUT)s $(ARGS) $(TESTPKGS)
 
-test-xml: fmt lint vendor | $(BASE) $(GO2XUNIT) ; $(info $(M) running $(NAME:%=% )tests…) @ ## Run tests with xUnit output
+test-xml: fmt lint | $(BASE) $(GO2XUNIT) ; $(info $(M) running $(NAME:%=% )tests…) @ ## Run tests with xUnit output
 	$Q cd $(BASE) && 2>&1 $(GO) test -timeout 20s -v $(TESTPKGS) | tee test/tests.output
 	$(GO2XUNIT) -fail -input test/tests.output -output test/tests.xml
 
@@ -107,7 +75,7 @@ COVERAGE_HTML = $(COVERAGE_DIR)/index.html
 .PHONY: test-coverage test-coverage-tools
 test-coverage-tools: | $(GOCOVMERGE) $(GOCOV) $(GOCOVXML)
 test-coverage: COVERAGE_DIR := $(CURDIR)/test/coverage.$(shell date -u +"%Y-%m-%dT%H:%M:%SZ")
-test-coverage: fmt lint vendor test-coverage-tools | $(BASE) ; $(info $(M) running coverage tests…) @ ## Run coverage tests
+test-coverage: fmt lint test-coverage-tools | $(BASE) ; $(info $(M) running coverage tests…) @ ## Run coverage tests
 	$Q mkdir -p $(COVERAGE_DIR)/coverage
 	$Q cd $(BASE) && for pkg in $(TESTPKGS); do \
 	  $(GO) test \
@@ -121,35 +89,12 @@ test-coverage: fmt lint vendor test-coverage-tools | $(BASE) ; $(info $(M) runni
 	$Q $(GO) tool cover -html=$(COVERAGE_PROFILE) -o $(COVERAGE_HTML)
 	$Q $(GOCOV) convert $(COVERAGE_PROFILE) | $(GOCOVXML) > $(COVERAGE_XML)
 
-.PHONY: lint
-lint: vendor | $(BASE) $(GOLINT) ; $(info $(M) running golint…) @ ## Run golint
-	$Q cd $(BASE) && ret=0 && for pkg in $(LINT_PKGS); do \
-	  test -z "$$($(GOLINT) $$pkg | tee /dev/stderr)" || ret=1 ; \
-	 done ; exit $$ret
 
 .PHONY: fmt
 fmt: ; $(info $(M) running gofmt…) @ ## Run gofmt on all source files
 	@ret=0 && for d in $$($(GO) list -f '{{.Dir}}' ./... | grep -v /vendor/); do \
 	  $(GOFMT) -l -w $$d/*.go || ret=$$? ; \
 	 done ; exit $$ret
-
-# Dependency management
-
-vendor: Gopkg.toml Gopkg.lock | $(BASE) $(GODEP) ; $(info $(M) retrieving dependencies…)
-	$Q cd $(BASE) && $(GODEP) ensure
-	@ln -nsf . vendor/src
-	@touch $@
-.PHONY: vendor-update
-vendor-update: | $(BASE) $(GODEP)
-ifeq "$(origin PKG)" "command line"
-	$(info $(M) updating $(PKG) dependency…)
-	$Q cd $(BASE) && $(GODEP) ensure -update $(PKG)
-else
-	$(info $(M) updating all dependencies…)
-	$Q cd $(BASE) && $(GODEP) ensure -update
-endif
-	@ln -nsf . vendor/src
-	@touch vendor
 
 # Misc
