+++ go.mod
@@ -1,0 +1,12 @@
+module github.com/byxorna/homer
+
+go 1.15
+
+require (
+	github.com/miekg/dns v1.0.3
+	github.com/patrickmn/go-cache v2.1.0+incompatible
+	golang.org/x/crypto v0.0.0-20191011191535-87dc89f01550
+	golang.org/x/lint v0.0.0-20201208152925-83fdc39ff7b5 // indirect
+	golang.org/x/net v0.0.0-20190620200207-3b0461eec859
+	gopkg.in/yaml.v2 v2.0.0
+)
